Masto tweaks (#6)
* Turn on keepCw by default

In Misskey, CW's are not automatically included in replies to a CW'ed
post unless the keepCw registry flag is set. This flag defaults to false
and there is not an easy user toggle for it.

To make it easier for users moving to Misskey from Mastodon, this
change enables the registry flag by default.

* add mention for all replies based on registry flag

Misskey by default only puts in username mentions when replying
to users off instance. This looks strange when viewing the posts on
mastodon. This patch includes said mentions even for on-instance
users.

This change originally came from a patch file, but subsequent
changes kept the patch from being applied. The relevant sections
of code have been modified here by me. The original patch is from
Cal <cal@noreply.clockworkcaracal.com>.

This patch was also modified to be conditional on a registry flag
that defaults to true.