Repo cleanup
Autoformatters had caused a significant number of changes to the
source files, despite no real change in functionality. This caused
merge tools to fail, requiring manual merge, as well as increasing
the chance of added bugs. This commit takes all "trivial"
differences and brings them in sync with upstream. It also
resolves a handul of cases where an upstream patch was applied
incorrectly, leading to cases of frankensteinian code.