Increase bio character limit
The default character limit for user profile bios is too low, and both
limits local users and can prevent some remote users from federating.
This change increases the limit to 2^14 (in the ORM, for remote users)
and 16k (for local users).

Note that this currently requires manual database schema editing to take
effect.