Allow restricting access to the client API
The client API, used for serving public HTML pages both to authenticated
and unauthenticated users, can be used by people off instance to access
data on our users. This commmit allows this API to be restricted on
startup using the flag privateClientApi in the config file.

This is accomplished by having endpoints refer to the config file when
specifying the requireCredential flag.