Verify signatures on incoming AP requests
When the requireSignToActivityPubGet config option is set to true,
misskey will require ActivityPub requests to have a valid signature
attached. This signature will then be checked against the list of
blocked hosts.

The end result of this is that only instances we federate with are able
to retreive content posted on our instance.