Verify signatures on incoming AP requests
When the secureMode config option is set, Misskey will require
ActivityPub GET requests to have a valid HTTP signature attached. This
signature will then be checked against the list of blocked hosts.

The end result of this is that only instances we federate with are able
to retrieve content posted on our instance.