Add Pluskey settings page
While we had made some tweaks to client behavior in the past, we did not
have an easy way for users to enable or disable said behavior besides
manual registry manipulation. This commit adds a page to the settings
menu for Pluskey settings and populates the locale files with relevant
strings.

Note that because I cannot translate to all languages myself, the
settings page is currently only in English.