Disable instance overview page by default
When the instance overview page is opened, the Misskey client makes a
number of expensive API calls for graphs. On low power or single core
servers, this can lock up the server for several seconds while these
graphs are prepared.

Furthermore, the instance overview page is opened by default whenever an
admin opens the "instance" menu. Most of the time, admins are trying to
open up a different submenu, and do not care or look at the instance
overview when it loads. This effectively wastes large portions of the
server's time and makes it unresponsive for no reason.

This commit stops the overview page from appearing when opening the
instance menu. Admins can still access this page if they choose, but it
requires specifically opening up the overview page.