Make notes optionally clickable
This commit adds the "noteLink" client registry setting. When this
setting is enabled, the user can click on the note to be directed to the
note's page.

It's non-obvious that clicking the timestamp on a note in Misskey will
bring up the detailed note page, and new users often ask how to browse
threads. Additionally, browsing threads by clicking through timestamps
can be somewhat painful from a usability standpoint.

With the "noteLink" setting enabled, a link is put underneath the note's
content. This allows the user to click on the negative space within a
note to be directed straight to the note thread's page.

This also applies to subnotes, so the user can browse through threads by
clicking on the negative space in any note in the thread.

Because this option is somewhat experimental, has the potential to break
CSS in some cases, and has not been tested for accessibility, it is
off by default and the setting is placed under a "Experimental" category
within the Pluskey settings page.