Optionally ignore repeated renotes
When the "skipRepeatRenotes" setting is enabled, if a note is renoted
several times in short succession only one copy will be displayed in the
client.

One common annoyance is that the Misskey timeline is too hectic, as if a
post blows up in renotes there can often be many copies of it on screen
at once.

This commit adds the "skipRepeatRenotes" registry setting, with a toggle
under the "experimental" section of the Pluskey settings page. When this
flag is set, the client will only put one copy of a post on the timeline
at once.

This also removes the sound effects for said post appearing,
unless it is caused by the user themselves renoting a post. Quote
renotes are not treated as renotes for the purpose of this check.

This flag is under the "experimental" section because it does not
entirely work. When first opening a timeline, there is a bug where all
renotes will be displayed, even if they are duplicates. However, as
posts stream in through the streaming endpoint, they appear to be
filtered correctly. Because the primary annoyance with these involves
posts that are streaming in, I decided to include this patch in anyway.